Deepchem_toxicitymodel
This is an introductory deep learning model, which uses the Deepchem package. A lots of credit has to be given to the authors(Ramsundar B., Eastman P., Walters P., Pande V.) of the book "Deep Learning for the Life Sciences". This book has been a gold mine in terms of my exploration and journey in DeepLearning, especially in the field of biochemistry.